+++ src/job.js
@@ -259,18 +259,58 @@ class Job extends BaseEntity {
 	 * @async
 	 * @returns {Promise<object.<string, *>>} The JSON-based response compatible to the API specification, but also including a `costs` property if present in the headers.
 	 * @throws {Error}
+	 * @deprecated
 	 */
 	async getResultsAsItem() {
+		let data = await this.getResultsAsStac();
+		if (data.type === 'Feature') { // Item
+			return data;
+		}
+		else { // We got a Collection, try to make a minimal Item from it for backward-compatibility
+			let item = Utils.pickFromObject(data, ['stac_version', 'id', 'assets', 'links']);
+			item.type = 'Feature';
+			item.geometry = null;
+			item.properties = Utils.pickFromObject(data, ['title', 'description', 'license', 'providers', 'created', 'updated', 'expires', 'costs']);
+			item.properties.datetime = null;
+			if (Utils.isObject(data.extent) && Utils.isObject(data.extent.temporal) && Array.isArray(data.extent.temporal.interval) && Array.isArray(data.extent.temporal.interval[0])) {
+				let temp = data.extent.temporal.interval[0];
+				if (typeof temp[0] === 'string' && temp[1] === 'string') {
+					item.properties.start_datetime = temp[0];
+					item.properties.end_datetime = temp[1];
+				}
+				else {
+					item.properties.datetime = temp[0] || temp[1];
+				}
+			}
+			return item;
+		}
+	}
+
+	/**
+	 * Retrieves the STAC Item or Collection produced for the job results.
+	 * 
+	 * @async
+	 * @returns {Promise<object.<string, *>>} The JSON-based response compatible to the API specification, but also including a `costs` property if present in the headers.
+	 * @throws {Error}
+	 */
+	async getResultsAsStac() {
 		let response = await this.connection._get('/jobs/' + this.id + '/results');
 		let data = response.data;
-		if (!Utils.isObject(data.properties)) {
-			data.properties = {};
-		}
 		if (!Utils.isObject(data.assets)) {
 			data.assets = {};
 		}
-		if (typeof response.headers['openeo-costs'] === 'number') {
-			data.properties.costs = response.headers['openeo-costs'];
+		if (data.type === 'Feature') { // Item
+			if (!Utils.isObject(data.properties)) {
+				data.properties = {};
+			}
+			if (typeof response.headers['openeo-costs'] === 'number') {
+				data.properties.costs = response.headers['openeo-costs'];
+			}
+		}
+		else { // Collection
+			if (typeof response.headers['openeo-costs'] === 'number') {
+				data.costs = response.headers['openeo-costs'];
+			}
 		}
 
 		return data;
@@ -284,7 +324,7 @@ class Job extends BaseEntity {
 	 * @throws {Error}
 	 */
 	async listResults() {
-		let item = await this.getResultsAsItem();
+		let item = await this.getResultsAsStac();
 		if (Utils.isObject(item.assets)) {
 			return Object.values(item.assets);
 		}